I believe I know what the fuckup was
I had my optimization flag set to -O3, which I believe can cause some wonky behaviour when some code is misaligned. Don't quote me on that.
I'll set it to -O2 for now, and see what gives, and at the end if I still have time and energy for it, I'll do some optimization of my own to make -O3 work

Forgot to mention that one of the fixes that I did for my previous bug of misadressed function pointers
was to malloc my main container, which fixed some of the issues.

The logic now works, just need to properly update the biggest/smallest position iterators on push in a subfunction, and we're good to actually start the sorting process, lol