Think I finally have insert_b behaving like I want
Haven't done any testing yet, but the refactoring is done, with a lot less unnecessary calculations.

Still A LOT of trimming to do on insert_elem_b (could easily be 5-10 lines with two subfcts), but not the goal for now

Tomorrow we start testing, and if it all works I'm considering myself good to move on to bigger sorting algos.

In principle, this algo sorts all possibilities for 5 in 8 or less moves, with only 4-5 that reach 8