Phew, tired
The last optimization I've done hasn't resulted in any improvement, and I think it's because I'm accidentally increasing the disorder. The use case for rr is wrong.

next_div should start at the biggest and then the smallest etc.
Worth a shot

Nah, no luck down that road.

I think my next best chance is to use insert_a for the first half, then insertion sort for the rest.
Reasoning is that the real bottleneck of my algo is inserting the last half b/c of high disorder.
If I can already have that half sorted, then I'm freeing the bottleneck.

The idea is that the high number of values in stack_a should not hinder too badly the insert_a algo, as values in b will be free to rotate.

There should definitely be a threshold of distance, b/c *some* disorder is acceptable I guess